fully replace pathLogic with pathsDataLogic
