+++ frontend/src/scenes/paths/renderPaths.ts
@@ -33,7 +33,7 @@ const createSankey = (width: number, height: number): Sankey.SankeyLayout<any, a
 const appendPathNodes = (
     svg: any,
     nodes: PathNodeData[],
-    filter: Partial<PathsFilterType>,
+    pathsFilter: Partial<PathsFilterType>,
     setNodeCards: Dispatch<SetStateAction<PathNodeData[]>>
 ): void => {
     svg.append('g')
@@ -62,7 +62,7 @@ const appendPathNodes = (
                     }
                 }
             }
-            if (isSelectedPathStartOrEnd(filter, d)) {
+            if (isSelectedPathStartOrEnd(pathsFilter, d)) {
                 return d3.color('purple')
             }
             const startNodeColor = c && d3.color(c) ? d3.color(c) : d3.color('#5375ff')
@@ -200,7 +200,7 @@ export function renderPaths(
     canvasWidth: number,
     canvasHeight: number,
     paths: { links: PathNode[]; nodes: any[] },
-    filter: Partial<PathsFilterType>,
+    pathsFilter: Partial<PathsFilterType>,
     setNodeCards: Dispatch<SetStateAction<PathNodeData[]>>
 ): void {
     if (!paths || paths.nodes.length === 0) {
@@ -227,7 +227,7 @@ export function renderPaths(
 
     setNodeCards(nodes.map((node: PathNodeData) => ({ ...node, visible: node.y1 - node.y0 > HIDE_PATH_CARD_HEIGHT })))
 
-    appendPathNodes(svg, nodes, filter, setNodeCards)
+    appendPathNodes(svg, nodes, pathsFilter, setNodeCards)
     appendDropoffs(svg)
     appendPathLinks(svg, links, nodes, setNodeCards)
     addChartAxisLines(svg, height, nodes, maxLayer)
